replace the big decimal with double

+++ src/entity/Money.kt
@@ -1,9 +1,8 @@
 package entity
 
-import java.math.BigDecimal
 
 data class Money(
-    val amount: BigDecimal,
+    val amount: Double,
     val currencyType: Currency = Currency.USD
 ){
 enum class Currency {
